Moved ERC20InfoCache.loadXXX() to ERCUtils.
